+++ apps/block_scout_web/test/block_scout_web/controllers/address_controller_test.exs
@@ -56,21 +56,10 @@ defmodule BlockScoutWeb.AddressControllerTest do
   describe "GET show/3" do
     setup :set_mox_global
 
-    setup do
-      configuration = Application.get_env(:explorer, :checksum_function)
-      Application.put_env(:explorer, :checksum_function, :eth)
-
-      :ok
-
-      on_exit(fn ->
-        Application.put_env(:explorer, :checksum_function, configuration)
-      end)
-    end
-
     test "redirects to address/:address_id/transactions", %{conn: conn} do
       insert(:address, hash: "0x5aaeb6053f3e94c9b9a09f33669435e7ef1beaed")
 
-      conn = get(conn, "/address/0x5aAeb6053F3E94C9b9A09f33669435E7Ef1BeAed")
+      conn = get(conn, "/address/#{Address.checksum("0x5aAeb6053F3E94C9b9A09f33669435E7Ef1BeAed")}")
 
       assert html_response(conn, 200)
     end