+++ apps/explorer/lib/explorer/migrator/reindex_internal_transactions_with_incompatible_status.ex
@@ -57,6 +57,7 @@ defmodule Explorer.Migrator.ReindexInternalTransactionsWithIncompatibleStatus do
       t in Transaction,
       as: :transaction,
       where: t.status == ^:error,
+      where: not is_nil(t.block_number),
       where: not exists(pbo_query),
       where: exists(it_query),
       where: not exists(it_error_query)