+++ apps/indexer/lib/indexer/block/catchup/bound_interval_supervisor.ex
@@ -68,7 +68,8 @@ defmodule Indexer.Block.Catchup.BoundIntervalSupervisor do
 
     block_interval = Map.get(named_arguments, :block_interval, @block_interval)
     minimum_interval = div(block_interval, 2)
-    bound_interval = BoundInterval.within(minimum_interval..(minimum_interval * 10))
+    maximum_interval = (minimum_interval + :timer.seconds(1)) * 10
+    bound_interval = BoundInterval.within(minimum_interval..maximum_interval)
 
     %__MODULE__{
       fetcher: %Catchup.Fetcher{block_fetcher: block_fetcher, memory_monitor: Map.get(named_arguments, :memory_monitor)},